janus: dlclose() all loaded plugins on shutdown
This fixes a memory leak. However, we cannot do this unconditionally,
because calling dlclose() unloads the debug symbols needed by gdb and
valgrind (see: https://bugs.kde.org/show_bug.cgi?id=79362). As a
compromise, call dlclose() when log_level is low, which is likely in a
production environment. When debugging, do not call dlclose(). A
suppression for the leak already exists in janus.supp.